HDPI-772 - Separating the tasks into different components
