+++ src/main/java/uk/gov/hmcts/reform/pcs/notify/service/NotificationService.java
@@ -4,7 +4,7 @@ import com.github.kagkarlsson.scheduler.SchedulerClient;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
-import uk.gov.hmcts.reform.pcs.notify.config.EmailTaskConfiguration;
+import uk.gov.hmcts.reform.pcs.notify.config.SendEmailTaskComponent;
 import uk.gov.hmcts.reform.pcs.notify.domain.CaseNotification;
 import uk.gov.hmcts.reform.pcs.notify.exception.NotificationException;
 import uk.gov.hmcts.reform.pcs.notify.model.EmailNotificationRequest;
@@ -62,7 +62,7 @@ public class NotificationService {
         );
 
         boolean scheduled = schedulerClient.scheduleIfNotExists(
-            EmailTaskConfiguration.sendEmailTask
+            SendEmailTaskComponent.sendEmailTask
                 .instance(taskId)
                 .data(emailState)
                 .scheduledTo(Instant.now())
